+++ src/types.jl
@@ -146,6 +146,54 @@ end
 Base.copy(N::NotEqualTo) = NotEqualTo(N.value)
 
 #====================================================================================
+====================== TYPES FOR TRAVERSING ========================================
+====================================================================================#
+
+abstract type Priority end
+
+struct PriorityDFS{T<:Real} <: Priority
+    depth::Int
+    bound::T
+    neg_idx::Int # the negative backtrack index (negative because of maximizing)
+end
+
+function Base.isless(p1::PriorityDFS, p2::PriorityDFS)
+    if p1.depth < p2.depth
+        return true
+    elseif p1.depth == p2.depth
+        if p1.bound < p2.bound
+            return true
+        elseif p1.bound == p2.bound
+            return p1.neg_idx < p2.neg_idx
+        else
+            return false
+        end
+    end
+    return false
+end
+
+struct PriorityBFS{T<:Real} <: Priority
+    bound::T
+    depth::Int
+    neg_idx::Int # the negative backtrack index (negative because of maximizing)
+end
+
+function Base.isless(p1::PriorityBFS, p2::PriorityBFS)
+    if p1.bound < p2.bound
+        return true
+    elseif p1.bound == p2.bound
+        if p1.depth < p2.depth
+            return true
+        elseif p1.depth == p2.depth
+            return p1.neg_idx < p2.neg_idx
+        else
+            return false
+        end
+    end
+    return false
+end
+
+#====================================================================================
 ====================== TYPES FOR CONSTRAINTS ========================================
 ====================================================================================#
 
@@ -173,6 +221,14 @@ mutable struct MatchingInit
     parents::Vector{Int}
     used_l::Vector{Bool}
     used_r::Vector{Bool}
+end
+
+mutable struct SCCInit
+    index_ei::Vector{Int}
+    ids::Vector{Int}
+    low::Vector{Int}
+    on_stack::Vector{Bool}
+    group_id::Vector{Int}
 end
 
 """
@@ -277,6 +333,7 @@ mutable struct AllDifferentConstraint <: Constraint
     di_ei::Vector{Int}
     di_ej::Vector{Int}
     matching_init::MatchingInit
+    scc_init::SCCInit
     # corresponds to `in_all_different`: Saves the constraint idxs where all variables are part of this alldifferent constraint
     sub_constraint_idxs::Vector{Int}
 end
@@ -444,6 +501,7 @@ mutable struct ConstraintSolverModel{T<:Real}
     c_backtrack_idx::Int
     c_step_nr::Int
     backtrack_vec::Vector{BacktrackObj{T}}
+    backtrack_pq::PriorityQueue
     sense::MOI.OptimizationSense
     objective::ObjectiveFunction
     var_in_obj::Vector{Bool} # saves whether a variable is part of the objective function