+++ test/fcts.jl
@@ -517,35 +517,49 @@
 
     @testset "Traverse" begin
         com = CS.ConstraintSolverModel()
+        com.options.traverse_strategy = :DFS
         com.traverse_strategy = Val(:DFS)
         com.sense = MOI.MIN_SENSE
-        backtrack_vec = Vector{CS.BacktrackObj{Float64}}()
-        bounds = [0.4, 0.15, 0.15, 0.1]
-        depths = [3, 2, 1, 2]
+        com.backtrack_vec = Vector{CS.BacktrackObj{Float64}}()
+        bounds = [0.4, 0.15, 0.15, 0.1, 0.1]
+        depths = [3, 2, 1, 2, 2]
         bo = CS.BacktrackObj(com)
         for i = 1:length(bounds)
             bo.status = :Open
             bo.idx = i
             bo.best_bound = bounds[i]
             bo.depth = depths[i]
-            push!(backtrack_vec, bo)
+            push!(com.backtrack_vec, bo)
+            CS.add2priorityqueue(com, com.backtrack_vec[end])
         end
-        order = [1,4,2,3]
+        order = [1,4,5,2,3]
         for i=1:length(bounds)
-            found, bo = CS.get_next_node(com, backtrack_vec, true)
+            found, bo = CS.get_next_node(com, com.backtrack_vec, true)
             @test found
             @test bo.idx == order[i]
-            bo.status = :Closed
+            CS.close_node!(com, bo.idx)
         end
 
-        # test Best first search
+        com = CS.ConstraintSolverModel()
         com.traverse_strategy = Val(:BFS)
-        for i=1:length(bounds)
-            backtrack_vec[i].status = :Open
+        com.options.traverse_strategy = :BFS
+        com.sense = MOI.MIN_SENSE
+        com.backtrack_vec = Vector{CS.BacktrackObj{Float64}}()
+        bounds = [0.4, 0.15, 0.15, 0.1, 0.1]
+        depths = [3, 2, 1, 2, 2]
+        bo = CS.BacktrackObj(com)
+        for i = 1:length(bounds)
+            bo.status = :Open
+            bo.idx = i
+            bo.best_bound = bounds[i]
+            bo.depth = depths[i]
+            push!(com.backtrack_vec, bo)
+            CS.add2priorityqueue(com, com.backtrack_vec[end])
         end
-        order = [4,2,3,1]
+
+        order = [4,5,2,3,1]
         for i=1:length(bounds)
-            found, bo = CS.get_next_node(com, backtrack_vec, true)
+            found, bo = CS.get_next_node(com, com.backtrack_vec, true)
             @test found
             @test bo.idx == order[i]
             bo.status = :Closed