+++ test/runtests.jl
@@ -48,6 +48,7 @@ include("moi.jl")
 include("constraints/table.jl")
 include("constraints/indicator.jl")
 include("constraints/reified.jl")
+include("constraints/equal_to.jl")
 
 include("lp_solver.jl")